+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change `renovate` configuration to only get upgrades for CAPA compatible versions of k8s dependencies.
+- Don't reconcile the `ShareReconciler` if only the status field has changed.
 
 ## [0.19.0] - 2025-06-17